fix: renamed default conf file and relocated it
