+++ lua/chadrc.lua
@@ -97,6 +97,6 @@ local M = {
     }
 }
 
-M = vim.tbl_deep_extend("force", require("config.config"), M)
+M = vim.tbl_deep_extend("force", require("default_config"), M)
 
 return M
